Move test in another file
